% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-biobaby.Rmd.
\name{export_results}
\alias{export_results}
\title{导出MR分析结果到Excel文件}
\usage{
export_results(final_results, output_file = "results.xlsx")
}
\arguments{
\item{final_results}{MR分析的最终结果，数据框。}

\item{output_file}{字符串，输出Excel文件的文件名，默认为"results.xlsx"。}
}
\value{
无返回值。结果会保存到指定的Excel文件中。
}
\description{
此函数将MR分析的最终结果导出到指定的Excel文件中，包含两个工作表：最终结果和清理后的暴露数据。
}
\details{
此函数创建一个新的Excel工作簿，并将MR分析的最终结果和清理后的暴露数据分别写入两个工作表中。最终结果会保存到指定的Excel文件中。
}
\note{
请确保输入的数据框格式正确。
}
\examples{
export_results(final_results, output_file = "results.xlsx")
}
\keyword{Excel}
\keyword{export}
\keyword{results}