% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-biobaby.Rmd.
\name{opengwas_query}
\alias{opengwas_query}
\title{分片查询并缓存结果到本地文件}
\usage{
opengwas_query(snps, outcomes, chunk_size = 1000, cache_dir = "cache")
}
\arguments{
\item{snps}{字符向量，表示要查询的SNP字符串或列表，比如ukb-e-30800_CSA。}

\item{outcomes}{字符向量，表示要查询的结局列表。}

\item{chunk_size}{整数，每个分片的SNP数量，默认为1000。}

\item{cache_dir}{字符串，缓存文件的存储目录，默认为"cache"。}
}
\value{
所有分片查询的合并结果，数据框。
}
\description{
此函数将SNP列表分片查询API，并将结果缓存到本地文件，以避免重复查询。
}
\details{
此函数首先将SNP列表分片查询API，如果查询结果已缓存，则使用缓存结果，否则进行API查询并缓存结果。最后返回所有分片查询的合并结果。
}
\note{
请确保输入的SNP和结局列表是有效的。
}
\examples{
cache_dir <- "cache"
snps <- unique(exp_data$SNP[1:5000])
outcome <- opengwas_query(snps, "ukb-e-30800_CSA", cache_dir = cache_dir)
}
\keyword{api}
\keyword{cache}
\keyword{gwas}
\keyword{query}