% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-biobaby.Rmd.
\name{parallel_mr}
\alias{parallel_mr}
\title{执行MR分析}
\usage{
parallel_mr(exp_data, outcome_data, cores = availableCores()/2)
}
\arguments{
\item{exp_data}{暴露数据，数据框。}

\item{outcome_data}{结局数据，数据框。}

\item{cores}{整数，用于平行计算的核心数，默认为可用核心数的一半。}
}
\value{
MR分析的最终结果，数据框。
}
\description{
此函数执行暴露和结局数据的MR分析，包括数据和谐化，平行计算，以及MR分析结果的评估和过滤。
}
\details{
此函数首先和谐化暴露和结局数据，然后进行MR分析，并对结果进行评估和过滤，返回最终的MR分析结果。
}
\note{
请确保输入数据框的格式正确。
}
\examples{
final_results <- parallel_mr(exp_data, outcome_data, cores = 4)
}
\keyword{MR}
\keyword{analysis}
\keyword{harmonization}
\keyword{parallel}
\keyword{processing}