% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-biobaby.Rmd.
\name{process_data}
\alias{process_data}
\title{处理并合并暴露和结局数据}
\usage{
process_data(
  exp_data,
  ao_data,
  remove_eqtl = TRUE,
  filter_population = "European"
)
}
\arguments{
\item{exp_data}{暴露数据，数据框。}

\item{ao_data}{全部结局数据，数据框。}

\item{remove_eqtl}{逻辑值，是否移除暴露数据中的eQTL，默认为TRUE。}

\item{filter_population}{字符串，指定结局数据中需要筛选的人群，默认为"European"。}
}
\value{
合并后的筛选数据，数据框。
}
\description{
此函数处理暴露数据和结局数据，并根据指定条件进行筛选和合并。
}
\details{
此函数首先处理暴露数据，然后处理结局数据，最后合并筛选后的数据。
}
\note{
请确保输入数据框的格式正确。
}
\examples{
result_data <- process_data(exp_data, ao_data, remove_eqtl = TRUE, filter_population = "European")
}
\keyword{data}
\keyword{merge}
\keyword{processing}